refactor: 알람을 설정하는 메소드 변경
