+++ app/src/main/java/woowacourse/movie/model/movie/AlarmItem.kt
@@ -1,0 +1,9 @@
+package woowacourse.movie.model.movie
+
+import java.time.LocalDateTime
+
+data class AlarmItem(
+    val dateTime: LocalDateTime,
+    val title: String,
+    val subTitle: String,
+)
